feat(referral): move new program to /referral-program; keep old /referral
Gate the new referral program behind an internal-only route while the old page
stays live for everyone, until we open the new program to the public.

- New design now lives at /referral-program (ReferralProgram screen), registered
  in the tabs layout with href: null (direct link only, not in the tab bar).
- /referral is restored to the original points-based "Share your referral code"
  page; the Refer & Earn entry-point banners are reverted to their original copy
  and continue to point at /referral.
- Swap the how-it-works icons: card for the first row ("Have a Solid card
  issued."), wallet for the third ("Earn ...").
- Regenerate the asset registry for the updated referral images.
