+++ app/(protected)/(tabs)/_layout.tsx
@@ -190,6 +190,13 @@ export default function TabLayout() {
         }}
       />
       <Tabs.Screen
+        name="referral-program"
+        options={{
+          title: 'Referral Program',
+          href: null,
+        }}
+      />
+      <Tabs.Screen
         name="settings"
         options={{
           title: 'Settings',
