+++ lib/assets.ts
@@ -120,7 +120,7 @@ export const ASSETS = {
   },
   'images/card-yellow-background.png': {
     module: require('@/assets/images/card-yellow-background.png'),
-    hash: '59cf014d',
+    hash: '72e0b615',
   },
   'images/card.png': { module: require('@/assets/images/card.png'), hash: '0fc3742c' },
   'images/card.tsx': { module: require('@/assets/images/card.tsx'), hash: '896e9355' },
@@ -443,7 +443,7 @@ export const ASSETS = {
   },
   'images/referral-3d.png': {
     module: require('@/assets/images/referral-3d.png'),
-    hash: '79bcfd00',
+    hash: '8477b533',
   },
   'images/referral_large.png': {
     module: require('@/assets/images/referral_large.png'),
@@ -522,7 +522,7 @@ export const ASSETS = {
   },
   'images/share-yellow-background.png': {
     module: require('@/assets/images/share-yellow-background.png'),
-    hash: '7776c134',
+    hash: 'd25eb3c2',
   },
   'images/sign-out.tsx': { module: require('@/assets/images/sign-out.tsx'), hash: 'deb2a605' },
   'images/solid-4x.png': { module: require('@/assets/images/solid-4x.png'), hash: 'f6579942' },
@@ -659,14 +659,14 @@ export const ASSETS = {
   'images/veda.png': { module: require('@/assets/images/veda.png'), hash: 'e41bfe57' },
   'images/wallet-yellow-background.png': {
     module: require('@/assets/images/wallet-yellow-background.png'),
-    hash: '98629c6c',
+    hash: '86120c72',
   },
   'images/wallet.tsx': { module: require('@/assets/images/wallet.tsx'), hash: '64b9ce19' },
   'images/wallet_connect.png': {
     module: require('@/assets/images/wallet_connect.png'),
     hash: '7816178c',
   },
-  'images/warning.png': { module: require('@/assets/images/warning.png'), hash: '0e016e69' },
+  'images/warning.png': { module: require('@/assets/images/warning.png'), hash: '81dd623d' },
   'images/welcome-card.png': {
     module: require('@/assets/images/welcome-card.png'),
     hash: '01502018',
